+++ src/ExternalConnections/PurchasePolicies.java
@@ -12,7 +12,7 @@ public class PurchasePolicies {
         this.payment = payment;
     }
 
-    public int tryToPurchase (int total,int deliveryDetails){
+    public int tryToPurchase (float total, int deliveryDetails){
         int answer =0;
 
         System.out.println("trying to purchase with total:" + total +" and delivery weight is" +deliveryDetails );
